+++ src/main/java/jduke/data/exception/JdukeException.java
@@ -1,3 +1,5 @@
+package jduke.data.exception;
+
 public class JdukeException extends Exception {
     public JdukeException(String message) {
         super(String.format("%s", message));
