+++ src/main/java/jduke/task/Deadline.java
@@ -1,10 +1,14 @@
+package jduke.task;
+
 import java.time.LocalDate;
 import java.time.LocalTime;
 import java.time.format.DateTimeFormatter;
-public class Event extends Task {
+
+public class Deadline extends Task {
     private LocalDate date;
     private LocalTime time;
-    public Event(String description, String timing) {
+
+    public Deadline(String description, String timing) {
         super(description);
         String[] timingParams = timing.split(" ");
         if (timingParams.length == 2) {
@@ -14,11 +18,12 @@ public class Event extends Task {
     }
     @Override
     public String getType() {
-        return "E";
+        return "D";
     }
+
     @Override
     public String getDescription() {
-        return String.format("%s (at: %s%s)",
+        return String.format("%s (by: %s%s)",
                 super.getDescription(),
                 this.date.format(DateTimeFormatter.ofPattern("MMM dd yyyy")),
                 (this.time != null ? this.time.format(DateTimeFormatter.ofPattern(" hh:mm a")) : "")
@@ -33,7 +38,7 @@ public class Event extends Task {
     @Override
     public String toStorageFormat() {
         return String.format(
-                "E | %s | %s | %s%s",
+                "D | %s | %s | %s%s",
                 (super.isCompleted() ? "1" : "0"),
                 super.getDescription(),
                 this.date.format(DateTimeFormatter.ofPattern("d/M/yyyy")),
