+++ src/main/java/jduke/task/Task.java
@@ -1,3 +1,5 @@
+package jduke.task;
+
 import java.time.LocalDate;
 
 public abstract class Task {
