+++ src/main/java/jduke/ui/Ui.java
@@ -1,3 +1,5 @@
+package jduke.ui;
+
 import java.io.InputStream;
 import java.io.PrintStream;
 import java.util.Scanner;
